1. Removed ResponseEntities from controller 2. All errors are processed in services now 3. Now using Liquibase for DB initiation 4. Switched to PostgresSQL from MySql 5. Some code improvisation for better reading (removed boilerplate and duplicated code, reduced nesting and etc.) 6. Fixed some DTO classes (private modifiers and wrapper classes) 7. Fixed repositories (removed tons of native queries there) 8. Switched to Slf4j logger 9. Fixed README.md (PostgresSQL instead of MySQL usage)

+++ src/main/java/searchengine/dto/search/SearchData.java
@@ -9,5 +9,5 @@ public class SearchData {
     private String uri;
     private String title;
     private String snippet;
-    private double relevance;
+    private Float relevance;
 }
