+++ src/main/java/searchengine/services/search/RankedLemmaOnPageComparator.java
@@ -6,9 +6,9 @@ public class RankedLemmaOnPageComparator implements Comparator<RankedLemmaOnPage
 
     @Override
     public int compare(RankedLemmaOnPage o1, RankedLemmaOnPage o2) {
-        int compare = -(o1.getRank() - o2.getRank());
+        float compare = -(o1.getRank() - o2.getRank());
         if (compare != 0) {
-            return compare;
+            return (int)compare;
         }
         return o1.getLemma().compareTo(o2.getLemma());
     }
